+++ pulsar/apps/__init__.py
@@ -16,32 +16,6 @@ __all__ = ['Application',
            'ApplicationMonitor',
            'get_application']
 
-APP_EVENTS = ('start', 'stop')
-
-def app_events(app):
-    for name in APP_EVENTS:
-        yield name, Deferred(description='%s.%s' % (app.name, name))
-        
-def fire_event(f):
-    name = f.__name__[3:]
-    if name not in APP_EVENTS:
-        raise ValueError()
-    
-    def fire(result, app):
-        if not is_failure(result):
-            result = app
-        else:
-            result.log()
-        events.fire(name, app)
-        app.events[name].callback(app)
-        
-    def _(*args, **kwargs):
-        app = args[0].params.app
-        safe_async(f, args, kwargs).add_both(lambda r: fire(r, app))
-    
-    _.__name__ = f.__name__
-    _.__doc__ = f.__doc__
-    return _
             
 def get_application(name):
     '''Invoked in the arbiter domain, this function will return
@@ -74,22 +48,24 @@ An :class:`Actor` for serving a pulsar :class:`Application`.
     @property
     def app(self):
         return self.params.app
-
+    
+    def io_poller(self):
+        return self.app.io_poller(self)
+        
     # Delegates Callbacks to the application
     def on_start(self):
-        app = self.app
-        self.app_handler = app.handler()
+        self.app_handler = self.app.handler()
         self.app.worker_start(self)
         try:
             self.cfg.worker_start(self)
         except:
             pass
 
+    def on_info(self, data):
+        return self.app.worker_info(self, data)
+        
     def on_stop(self):
-        return self.app.worker_stop(self)
-
-    def on_exit(self):
-        self.app.worker_exit(self)
+        self.app.worker_stop(self)
         try:
             self.cfg.worker_exit(self)
         except:
@@ -98,9 +74,6 @@ An :class:`Actor` for serving a pulsar :class:`Application`.
     def on_info(self, info):
         return self.app.on_info(self, info)
     
-    def on_event(self, fd, event):
-        return self.app.on_event(self, fd, event)
-
 
 class ApplicationMonitor(Monitor):
     '''A :class:`Monitor` for managing a pulsar :class:`Application`.
@@ -118,54 +91,41 @@ class ApplicationMonitor(Monitor):
         
     ############################################################################
     # Delegates Callbacks to the application
-    @fire_event
     def on_start(self):
         super(ApplicationMonitor, self).on_start()
-        app = self.app
-        app.monitor_start(self)
+        self.app.monitor_start(self)
         # If no workers available invoke the worker start method too
         if not self.cfg.workers:
-            self.app_handler = app.handler()
-            app.worker_start(self)
+            self.app.worker_start(self)
         else:
-            self.app_handler = app.monitor_handler()
+            self.app_handler = self.app.monitor_handler()
+        events.fire('ready', self.app)
+        self.app.local.on_start.callback(self.app)
 
-    @fire_event
-    def on_stop(self):
-        app = self.params.app
+    def on_info(self, data):
         if not self.cfg.workers:
-            yield app.worker_stop(self)
-        yield app.monitor_stop(self)
-        yield super(ApplicationMonitor, self).on_stop()
-
-    def on_exit(self):
-        app = self.params.app
-        app.monitor_exit(self)
-        try:
-            self.cfg.worker_exit(self)
-        except:
-            pass
+            return self.app.worker_info(self, data)
+        else:
+            return self.app.monitor_info(self, data)
+            
+    def on_stop(self):
+        if not self.cfg.workers:
+            self.app.worker_stop(self)
+        self.app.monitor_stop(self)
+        super(ApplicationMonitor, self).on_stop()
+        self.app.stop.callback(self.app)
     
     def monitor_task(self):
         self.app.monitor_task(self)
         
-    def on_event(self, fd, event):
-        return self.app.on_event(self, fd, event)
-    
     def actorparams(self):
         p = Monitor.actorparams(self)
-        app = self.params.app
+        app = self.app
         if self.cfg.concurrency == 'thread':
             app = pickle.loads(pickle.dumps(app))
         p.update({'app': app,
                   'name': '{0}-worker'.format(app.name)})
-        return app.actorparams(self, p)
-
-    def on_info(self, info):
-        if not self.cfg.workers:
-            return self.params.app.on_info(self, info)
-        else:
-            return info
+        return self.app.actorparams(self, p)
 
 
 class Application(pulsar.Pulsar):
@@ -289,10 +249,9 @@ These are the most important facts about a pulsar :class:`Application`
         monitor = None
         if actor and actor.is_arbiter():
             monitor = actor.monitors.get(self.name)
-        if monitor is None and (not actor or actor.is_arbiter()): 
-            # Add events
-            self.local.events = dict(app_events(self))
+        if monitor is None and (not actor or actor.is_arbiter()):
             self.cfg.on_start()
+            self.local.on_start = Deferred()
             self.configure_logging()
             events.fire('ready', self)
             arbiter = pulsar.arbiter(cfg=self.cfg.new_config())
@@ -300,13 +259,11 @@ These are the most important facts about a pulsar :class:`Application`
                 monitor = arbiter.add_monitor(ApplicationMonitor,
                                               self.name,
                                               app=self,
-                                              cfg=self.cfg,
-                                              ioqueue=self.ioqueue)
+                                              cfg=self.cfg)
                 self.cfg = monitor.cfg
                 if self.commands_set:
                     monitor.impl.commands_set.update(self.commands_set)
-        if self.events:
-            return self.events['start']
+        return self.local.on_start
 
     @property
     def app(self):
@@ -320,10 +277,6 @@ These are the most important facts about a pulsar :class:`Application`
     def name(self):
         '''Application name, It is unique and defines the application.'''
         return self._name
-
-    @property
-    def events(self):
-        return self.local.events
     
     def __repr__(self):
         return self.name
@@ -336,12 +289,6 @@ These are the most important facts about a pulsar :class:`Application`
         actor = get_actor()
         if actor:
             return actor.monitors.get(self.name)
-         
-    @property
-    def ioqueue(self):
-        if 'queue' not in self.local:
-            self.local.queue = self.get_ioqueue()
-        return self.local.queue
 
     def handler(self):
         '''Returns the callable application handler which is stored in
@@ -349,31 +296,10 @@ These are the most important facts about a pulsar :class:`Application`
 By default it returns the :attr:`Application.callable`.'''
         return self.callable
 
-    def request_instance(self, worker, fd, event):
-        '''Build a request class from a file descriptor *fd* and an *event*.
-The returned request instance is passed to the :meth:`handle_request`
-method.'''
-        return event
-
-    def handle_request(self, worker, request):
-        '''This is the main function which needs to be implemented
-by actual applications. It is called by the *worker* to handle
-a *request*.
-
-:parameter worker: the :class:`Worker` handling the request.
-:parameter request: an application specific request object.
-:rtype: It can be a generator, a :class:`Deferred` instance
-    or the actual response.'''
-        raise NotImplementedError()
-
-    def get_ioqueue(self):
-        '''Returns an I/O distributed queue for the application if one
-is needed. If a queue is returned, the application :class:`Worker`
-will have a :class:`EventLoop` instance based on the queue (via :class:`IOQueue`).
-
-By default it returns ``None``.'''
+    def io_poller(self, worker):
+        '''called by :meth:`Actor.io_pooler`.'''
         return None
-
+    
     def on_config_init(self, cfg, params):
         '''Callback when configuration is initialised but not yet loaded.
 This is a chance to add extra :ref:`config parameters <settings>` or remove
@@ -492,9 +418,6 @@ By default it returns ``None``.'''
     # MONITOR AND WORKER CALLBACKS
     def on_info(self, worker, data):
         return data
-
-    def on_event(self, worker, fd, events):
-        pass
     
     # WORKERS CALLBACKS
     def worker_start(self, worker):
@@ -502,13 +425,11 @@ By default it returns ``None``.'''
 :ref:`callback <actor-callbacks>` method.'''
         pass
 
+    def worker_info(self, worker, data):
+        return data
+    
     def worker_stop(self, worker):
         '''Called by the :class:`Worker` :meth:`pulsar.Actor.on_stop`
-:ref:`callback <actor-callbacks>` method.'''
-        pass
-
-    def worker_exit(self, worker):
-        '''Called by the :class:`Worker` :meth:`pulsar.Actor.on_exit`
 :ref:`callback <actor-callbacks>` method.'''
         pass
 
@@ -523,12 +444,11 @@ before a new :class:`Worker` is spawned.'''
 The application is now in the arbiter but has not yet started.'''
         pass
 
+    def monitor_info(self, worker, data):
+        return data
+    
     def monitor_stop(self, monitor):
         '''Callback by :class:`ApplicationMonitor` when stopping'''
-        pass
-
-    def monitor_exit(self, monitor):
-        '''Callback by :class:`ApplicationMonitor` at exit'''
         pass
 
     def monitor_task(self, monitor):