+++ pulsar/apps/tasks/__init__.py
@@ -181,6 +181,7 @@ import pulsar
 from pulsar import to_string, maybe_async_deco, get_actor
 from pulsar.utils.importer import import_modules, module_attribute
 
+from .queue import *
 from .exceptions import *
 from .task import *
 from .models import *
@@ -194,7 +195,7 @@ class TaskQueueFactory(pulsar.Setting):
     name = "task_queue_factory"
     section = "Task Consumer"
     flags = ["-q", "--task-queue"]
-    default = "pulsar.Queue"
+    default = "pulsar.apps.tasks.Queue"
     desc = """The task queue factory to use."""
 
     def get(self):
@@ -231,22 +232,58 @@ This type of application is served by :ref:`CPU bound workers <cpubound>`.'''
         self.concurrent_requests = set()
         super(CPUboundServer, self).__init__(*args, **kwargs)
         
+    def io_poller(self, worker):
+        self.local.queue = worker.params.ioqueue 
+        return IOQueue(self.ioqueue, self)
+    
+    def can_poll(self):
+        if self.local.can_poll:
+            return len(self.concurrent_requests) <= self.cfg.backlog
+    
     @property
     def concurrent_request(self):
         return len(self.concurrent_requests)
-
-    def get_ioqueue(self):
-        '''Return the distributed task queue which produces tasks to
-be consumed by the workers.'''
-        return self.cfg.task_queue_factory()
+    
+    @property
+    def ioqueue(self):
+        return self.local.queue
+    
+    def put(self, request):
+        '''Put a *request* into the :attr:`ioqueue` if available.'''
+        self.ioqueue.put(('request', request))
 
     def request_instance(self, request):
+        '''Build a request class from a *request*. By default it returns the
+request. This method is called by the :meth:`on_request` once a new
+request has been obtained from the :attr:`ioqueue`.'''
         return request
 
     def worker_start(self, worker):
-        '''Set up the cpu bound worker and register its file descriptor'''
+        # Set up the cpu bound worker by registering its file descriptor
+        # and enabling polling from the queue
         worker.requestloop.add_reader('request', self.on_request, worker)
-        
+        self.local.can_poll = True
+    
+    def monitor_info(self, worker, data):
+        tq = self.ioqueue
+        if tq is not None:
+            if isinstance(tq, Queue):
+                tqs = 'multiprocessing.Queue'
+            else:
+                tqs = str(tq)
+            try:
+                size = tq.qsize()
+            except NotImplementedError: #pragma    nocover
+                size = 0
+            data['queue'] = {'ioqueue': tqs, 'ioqueue_size': size}
+        return data
+    
+    def actorparams(self, monitor, params):
+        if 'queue' not in self.local:
+            self.local.queue = self.cfg.task_queue_factory()
+        params['ioqueue'] = self.ioqueue
+        return params
+    
     @maybe_async_deco
     def on_request(self, worker, request):
         request = self.request_instance(request)