+++ pulsar/apps/tasks/scheduler.py
@@ -150,7 +150,7 @@ and task scheduling.
         task = self._make_request(jobname, targs, tkwargs, **params)
         if task.needs_queuing():
             task._queued = True
-            monitor.put(task.serialize_for_queue())
+            monitor.app.put(task.serialize_for_queue())
         else:
             task._queued = False
             self.logger.debug('Task %s already requested, abort.', task)