+++ pulsar/async/concurrency.py
@@ -115,7 +115,7 @@ class ActorThread(ActorConcurrency, Thread):
         '''Called by the main thread to force termination.'''
         actor = get_actor_from_id(self.aid)
         if actor is not None:
-            actor.exit()
+            actor.stop(1)
 
     @property
     def pid(self):