+++ pulsar/utils/events/__init__.py
@@ -3,7 +3,11 @@
 Originally based on pydispatch (BSD) http://pypi.python.org/pypi/PyDispatcher/2.0.1
 See license.txt for original license.
 """
+import logging
 from .dispatcher import Signal
+
+
+LOGGER = logging.getLogger('pulsar.events')
 
 _events = {}
 
@@ -26,3 +30,27 @@ def fire(name, sender, **kwargs):
     '''Fire event *name* from *sender*.'''
     event = create(name)
     event.send(sender, **kwargs)
+    
+    
+class EventHandler(object):
+    EVENTS = ()
+    
+    def __new__(cls, *args, **kwargs):
+        o = super(EventHandler, cls).__new__(cls)
+        o.hooks = dict(((event, []) for event in cls.EVENTS))
+        return o
+        
+    def bind_event(self, event, hook):
+        '''Register an event hook'''
+        self.hooks[event].append(hook)
+        
+    def fire(self, event, *event_data):
+        """Dispatches an event dictionary on a given piece of data."""
+        hooks = self.hooks
+        if hooks and event in hooks:
+            for hook in hooks[event]:
+                try:
+                    hook(*event_data)
+                except Exception:
+                    LOGGER.exception('%s unhandled error in %s hook',
+                                     self, event)